move cpu_usage to utils; change cpu limits to floats
git-svn-id: svn+ssh://astrometry.net/svn/trunk/src/astrometry@8627 03a3cea6-2c03-0410-ac9b-9271b1c66c29

+++ blind/augment-xylist.h
@@ -67,7 +67,7 @@ struct augment_xylist_s {
     double scalehi;
     char* scaleunits;
 
-    int cpulimit;
+    float cpulimit;
 
     bool tweak;
     int tweakorder;
